split cell into separate file
